libtirpc: Fix "libtirpc\src\svc_vc.c(392,20): warning C4018: '>': signed/unsigned mismatch" warning in |rendezvous_request()|
Fix "libtirpc\src\svc_vc.c(392,20): warning C4018: '>': signed/unsigned
mismatch" warning in |rendezvous_request()|.

+++ libtirpc/src/svc_vc.c
@@ -389,8 +389,8 @@ again:
 		if (fcntl(sock, F_SETFL, flags | O_NONBLOCK) == -1)
 			return (FALSE);
 #endif /* _WINTIRPC */
-		if (cd->recvsize > cd->maxrec)
-			cd->recvsize = cd->maxrec;
+		if (cd->recvsize > (u_int)cd->maxrec)
+			cd->recvsize = (u_int)cd->maxrec;
 		cd->nonblock = TRUE;
 		__xdrrec_setnonblock(&cd->xdrs, cd->maxrec);
 	} else
